refactor: moved LogLevelEnum to a separate file and updated imports
