+++ src/modules/common/enums/logLevel.enum.ts
@@ -1,0 +1,10 @@
+
+export enum LogLevelEnum {
+	// NOTE - ordered by level priority
+	ERROR = 'error',
+	WARN = 'warn',
+	INFO = 'info',
+	HTTP = 'http',
+	VERBOSE = 'verbose',
+	DEBUG = 'debug',
+}
